requirements.txt
certifi==2018.4.16
chardet==3.0.4
client==0.0.1
Django==2.0.6
django-crispy-forms==1.7.2
docopt==0.6.2
idna==2.7
Pillow==5.1.0
pipreqs==0.4.9
psycopg2-binary==2.7.5
pytz==2018.4
requests==2.19.1
urllib3==1.23
yarg==0.1.9
